+++ src/MicroElements.Swashbuckle.FluentValidation/RuleContext.cs
@@ -2,97 +2,42 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 using System.Reflection;
 using FluentValidation;
 using FluentValidation.Validators;
-using MicroElements.Swashbuckle.FluentValidation.Generation;
 using Microsoft.OpenApi.Models;
 
 namespace MicroElements.Swashbuckle.FluentValidation
 {
-    public record SchemaGenerationContext
-    {
-        /// <summary>
-        /// Gets OpenApi schema.
-        /// </summary>
-        public OpenApiSchema Schema { get; init; }
-
-        /// <summary>
-        /// Schema type.
-        /// </summary>
-        public Type SchemaType { get; init; }
-
-        public ReflectionContext ReflectionContext { get; init; }
-
-        public IReadOnlyList<FluentValidationRule> Rules { get; init; }
-
-        public ISchemaGenerationOptions SchemaGenerationOptions { get; init; }
-
-        public ISchemaProvider<OpenApiSchema> SchemaProvider { get; init; }
-}
-
-    /// <summary>
-    /// Contains <see cref="PropertyRule"/> and additional info.
-    /// </summary>
-    public record ValidationRuleContext
-    {
-        /// <summary>
-        /// PropertyRule.
-        /// </summary>
-        public IValidationRule PropertyRule { get; init; }
-
-        /// <summary>
-        /// Flag indication whether the <see cref="PropertyRule"/> is the CollectionRule.
-        /// </summary>
-        public bool IsCollectionRule { get; init; }
-
-        /// <summary>
-        /// Reflection context.
-        /// </summary>
-        public ReflectionContext ReflectionContext { get; init; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ValidationRuleContext"/> class.
-        /// </summary>
-        /// <param name="propertyRule">PropertyRule.</param>
-        /// <param name="isCollectionRule">Is a CollectionPropertyRule.</param>
-        /// <param name="reflectionContext">Reflection context.</param>
-        public ValidationRuleContext(
-            IValidationRule propertyRule,
-            bool isCollectionRule,
-            ReflectionContext reflectionContext)
-        {
-            PropertyRule = propertyRule;
-            IsCollectionRule = isCollectionRule;
-            ReflectionContext = reflectionContext;
-        }
-    }
-
     /// <summary>
     /// RuleContext.
     /// </summary>
     public class RuleContext
     {
         /// <summary>
-        /// Swagger schema.
+        /// Gets OpenApi (swagger) schema.
         /// </summary>
         public OpenApiSchema Schema { get; }
 
         /// <summary>
-        /// Property name.
+        /// Gets property name in schema.
         /// </summary>
         public string PropertyKey { get; }
 
         /// <summary>
-        /// Property validator.
+        /// Gets <see cref="IValidationRule"/> with extended information.
+        /// </summary>
+        public ValidationRuleInfo ValidationRuleInfo { get; }
+
+        /// <summary>
+        /// Gets property validator for property in schema.
         /// </summary>
         public IPropertyValidator PropertyValidator { get; }
 
         /// <summary>
-        /// Gets value indicating that <see cref="PropertyValidator"/> should be applied to collection item instead of property.
+        /// Gets a value indicating whether the <see cref="PropertyValidator"/> should be applied to collection item instead of property.
         /// </summary>
-        public bool IsCollectionValidator { get; }
+        public bool IsCollectionValidator => ValidationRuleInfo.IsCollectionRule;
 
         /// <summary>
         /// Gets target property schema.
@@ -100,7 +45,7 @@ namespace MicroElements.Swashbuckle.FluentValidation
         public OpenApiSchema Property => !IsCollectionValidator ? Schema.Properties[PropertyKey] : Schema.Properties[PropertyKey].Items;
 
         /// <summary>
-        /// Reflection context.
+        /// Gets reflection context.
         /// </summary>
         public ReflectionContext ReflectionContext { get; }
 
@@ -109,32 +54,50 @@ namespace MicroElements.Swashbuckle.FluentValidation
         /// </summary>
         /// <param name="schema">Swagger schema.</param>
         /// <param name="propertyKey">Property name.</param>
+        /// <param name="validationRuleInfo">ValidationRuleInfo.</param>
         /// <param name="propertyValidator">Property validator.</param>
         /// <param name="reflectionContext">Reflection context.</param>
-        /// <param name="isCollectionValidator">Should be applied to collection items.</param>
         public RuleContext(
             OpenApiSchema schema,
             string propertyKey,
+            ValidationRuleInfo validationRuleInfo,
             IPropertyValidator propertyValidator,
-            ReflectionContext? reflectionContext = null,
-            bool isCollectionValidator = false)
+            ReflectionContext? reflectionContext = null)
         {
             Schema = schema;
             PropertyKey = propertyKey;
+            ValidationRuleInfo = validationRuleInfo;
             PropertyValidator = propertyValidator;
             ReflectionContext = reflectionContext;
-            IsCollectionValidator = isCollectionValidator;
         }
     }
 
+    /// <summary>
+    /// Reflection context for <see cref="RuleContext"/>.
+    /// </summary>
     public class ReflectionContext
     {
+        /// <summary>
+        /// Gets the type (schema type).
+        /// </summary>
         public Type? Type { get; }
 
+        /// <summary>
+        /// Gets optional PropertyInfo.
+        /// </summary>
         public MemberInfo? PropertyInfo { get; }
 
+        /// <summary>
+        /// Gets optional ParameterInfo.
+        /// </summary>
         public ParameterInfo? ParameterInfo { get; }
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ReflectionContext"/> class.
+        /// </summary>
+        /// <param name="type">Schema type.</param>
+        /// <param name="propertyInfo">Optional PropertyInfo.</param>
+        /// <param name="parameterInfo">Optional ParameterInfo.</param>
         public ReflectionContext(
             Type? type = null,
             MemberInfo? propertyInfo = null,
@@ -143,6 +106,13 @@ namespace MicroElements.Swashbuckle.FluentValidation
             Type = type;
             PropertyInfo = propertyInfo;
             ParameterInfo = parameterInfo;
+        }
+
+        public static ReflectionContext CreateFromProperty(MemberInfo propertyInfo)
+        {
+            return new ReflectionContext(
+                type: propertyInfo.ReflectedType,
+                propertyInfo: propertyInfo);
         }
     }
 }